# docs/swagger.yaml
basePath: /
definitions:
...
  dto.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      email:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
  dto.LogoutResponse:
    properties:
      msg:
        type: string
    type: object
...
  dto.Movies:
    properties:
      Id:
        type: integer
      background_url:
        type: string
      duration:
        type: integer
      poster_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      status:
        type: string
      synopsis:
        type: string
      title:
        type: string
    required:
    - Id
    type: object
  dto.NewUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.PaginationMeta:
    properties:
      next_page:
        type: string
      page:
        type: integer
      prev_page:
        type: string
      total_page:
        type: integer
    type: object
  dto.RegisterResponse:
    properties:
      email:
        type: string
    type: object
  dto.Response:
    properties:
      data:
        items: {}
        type: array
      error:
        type: string
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      msg:
        type: string
      success:
        type: boolean
    type: object
  dto.ResponseError:
    properties:
      error:
        type: string
      msg:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdateMovies:
    properties:
      background_url:
        type: string
      duration:
        type: integer
      poster_url:
        type: string
      rating:
        type: number
      release_date:
        type: string
      status:
        type: string
      synopsis:
        type: string
      title:
        type: string
    type: object
  dto.UpdatePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
host: localhost:8002
info:
  contact: {}
  description: Backend Services For Tickitz Project
  title: Tickitz Services
  version: "1.0"
paths:
...
  /admin/movies/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Movies'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Show All Movies For Admin
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Title
        in: formData
        name: title
        type: string
      - description: Synopsis
        in: formData
        name: synopsis
        type: string
      - description: Poster Photo
        in: formData
        name: poster_file
        type: file
      - description: Background Photo
        in: formData
        name: background_file
        type: file
      - description: Release Date
        in: formData
        name: release_date
        type: string
      - description: Duration Movie
        in: formData
        name: duration
        type: integer
      - description: Status(upcoming or now_showing)
        in: formData
        name: status
        type: string
      - description: Rating
        in: formData
        name: rating
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Post Movies
      tags:
      - admin
  /admin/movies/delete/{id}:
    delete:
      parameters:
      - description: Movie Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateMovies'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Delete Movies For Admin
      tags:
      - admin
  /admin/movies/update/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Movie Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Movies Body
        in: body
        name: movies
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMovies'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateMovies'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Update Movies For Admin
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Login Body
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Login new user
      tags:
      - auth
  /auth/logout:
    delete:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User Update Body
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdatePassword'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Registration Body
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/dto.NewUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Register new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"